aiString.length is really of type "ai_uint32" which corresponds to "c_uint32"
and not to "c_size_t" (which is different on 64bit systems and 32bit systems!)

Closes: https://github.com/assimp/assimp/issues/2788

+++ port/PyAssimp/pyassimp/structs.py
@@ -76,7 +76,7 @@ class String(Structure):
             # Binary length of the string excluding the terminal 0. This is NOT the
             #  logical length of strings containing UTF-8 multibyte sequences! It's
             #  the number of bytes from the beginning of the string to its end.
-            ("length", c_size_t),
+            ("length", c_uint32),
 
             # String buffer. Size limit is MAXLEN
             ("data", c_char*MAXLEN),
